pyproject.toml
[project]
name = "paper_theater01"
version = "0.1.0"
description = "紙芝居同期型テキストチャットシステム - ドライブナビゲーター"
requires-python = ">=3.10"
dependencies = [
    "gradio>=5.0.0,<6.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "fastrtc[vad]>=0.0.30",
    "scipy>=1.0.0",
    "numpy>=1.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "debugpy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["paper_theater01"]

[tool.uv]
dev-dependencies = []
